modularize app.py
separate out helper functions for routing and text processing
